requirements.txt
Django==1.6.6
Pillow==2.5.3
South==1.0
argparse==1.2.1
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-cors-headers==0.13
django-countries==2.1.2
django-guardian==1.2.4
django-sslify==0.2.3
django-toolbelt==0.0.1
django-userena==1.3.0
djangorestframework==2.3.14
easy-thumbnails==2.1
gunicorn==19.1.1
html2text==2014.7.3
jsonfield==1.0.0
newrelic==2.26.0.22
psycopg2==2.5.3
sendgrid==1.1.1
six==1.7.3
smtpapi==0.1.2
static3==0.5.1
unittest2==0.5.1
wsgiref==0.1.2
